requirements.txt
Flask==3.0.3
tensorflow==2.11.0
keras==2.11.0
numpy==1.23.5
Pillow==10.2.0
gunicorn==21.2.0
jinja2==3.1.4
protobuf==3.20.3
h5py==3.8.0